requirements.txt
google-cloud-aiplatform
google-cloud-bigquery
google-api-python-client
google-auth-httplib2
google-auth-oauthlib
pandas
Pillow
python-dotenv
requests